==> 0_1_get_paevakorrad_stenod.ipynb <==
import requests
import json
from copy import deepcopy
import pandas as pd
from bs4 import BeautifulSoup

## Helpers

def get_query_jsn(url):
    """returns json from request
    -INPUT:
        -url: str of url to be requested
    -OUTPUT:
        - list of päevakorrad"""
    return json.loads(requests.get(url).content)

def get_years_paevakorrad(years):
    """collects all päevakord from each year
    -years: list of integeres"""
    paevakorrad=[]
    for year in years:
        print(f'working on year {year}')
        try:
            paevakorrad_year=get_query_jsn(f'https://api.riigikogu.ee/api/agenda/plenary?endDate={year}-12-31&lang=et&querySteno=true&startDate={year}-01-01')
            paevakorrad.append(paevakorrad_year)
        except Exception as e:
            print(f'exception on year {year}, {e}')
            paevakorrad_year_first_half=get_query_jsn(f'https://api.riigikogu.ee/api/agenda/plenary?endDate={year}-06-01&lang=et&querySteno=true&startDate={year}-01-01')
            paevakorrad.append(paevakorrad_year_first_half)
            paevakorrad_year_second_half=get_query_jsn(f'https://api.riigikogu.ee/api/agenda/plenary?endDate={year}-12-31&lang=et&querySteno=true&startDate={year}-06-02')
            paevakorrad.append(paevakorrad_year_second_half)
    return paevakorrad

def paevakorrad_stenos_count(paevakorrad):
    """counts number of stenos 
    -INPUT:
        - paevakorrad: list of päevakorrad
    -OUTPUT:
        -pandas DataFrame with päevakorrad and stenos count info"""
    paevakorrad_stenos_count={}
    for year in paevakorrad:
        paevakorrad_stenos_count[year.get('weekStartDate')]={'agendaItems':0,
                                                            'have_stenos':0,
                                                            'stenos_links':[],
                                                            'stenos_n_unique_links':0}
        sittings=year.get('sittings', {})
        for sitting in sittings:
            agenda_items=sitting.get('agendaItems', {})
            for agenda_item in agenda_items:
                link=agenda_item.get('_links',{}).get('steno',{}).get('href', {})
                paevakorrad_stenos_count[year.get('weekStartDate', )]['agendaItems']+=1
                if link!={}:
                    paevakorrad_stenos_count[year.get('weekStartDate')]['have_stenos']+=1
                    paevakorrad_stenos_count[year.get('weekStartDate')]['stenos_links'].append(link.split('#')[0])
        paevakorrad_stenos_count[year.get('weekStartDate')]['stenos_n_unique_links']=len(set(paevakorrad_stenos_count[year.get('weekStartDate')]['stenos_links']))
    return pd.DataFrame(paevakorrad_stenos_count).T

def get_main_talk(soup, conversations=[], steno_link=None):
    """scrape content of stenogram
    -INPUT:
        - soup: beautifulsoup object of stenogram
        - conversations: list where to store stenograms
        - steno_link: str of url to steno
    -OUTPUT:
        lis of conversations (previous + currently parsed)"""
    for i, tag in enumerate(soup.find('article').next_siblings):
        if tag.name=='article':
            heading=tag.find('h3').text.strip()
            speech_areas=tag.findAll("div", {"class": "pb-4 speech-area"})
            for speech_area in speech_areas:
                link=speech_area.find('a', {'class':'steno-video-icon'})
                if link is not None:
                    link=link['href']
                else:
                    link=None

                speaker=speech_area.find('h4').text.strip()
                p_tags=speech_area.find_all('p')
                if p_tags is not None:
                    text=' '.join([p.get_text(separator=" ").strip() for p in p_tags])
                else:
                    text=None
                ind=tag['id']
                ind_snd=speech_area['id']
                conversation={'heading':heading,
                              'speaker':speaker,
                              'link_video':link,
                              'link_steno':steno_link,
                              'index_pk':ind,
                              'index_snd':ind_snd,
                             'text':text}
                conversations.append(conversation)
    return conversations

def get_agenda(soup, steno_link=None):
    """scrape steno agenda from soup
    -INPUT:
         - soup: beautifulsoup object of stenogram
         - steno_link: str of url to steno
    -OUTPUT:
        -list of conversation agenda information"""
    conversations=[]
    header=None
    text=None
    url=None
    for i, tag in enumerate(soup.find('article', {"class": "steno-agenda-item"})):
        if tag.name=='h3':
            heading=tag.text.strip()
        if tag.name=='div' and tag.attrs['class']!=['d-flex']:
            speaker=tag.find('h4').text.strip()
            p_tag=tag.find('p')
            if p_tag is not None:
                text=p_tag.get_text(separator=" ").strip()
            else:
                text=None
            href_tag=tag.find('a', {'class':'steno-video-icon'})
            if href_tag is not None:
                url=href_tag['href']
            else:
                url=None
            ind=tag.parent['id']
            ind_snd=tag['id']
            conversation={'heading':heading,
                              'speaker':speaker,
                              'link_video':url,
                              'link_steno':steno_link,
                              'index_pk':ind,
                              'index_snd':ind_snd,
                             'text':text}
            conversations.append(conversation)
            
    return conversations

def parse_steno(soup, steno_link):
    """wrapper function to get steno content and agenda
    -INPUT:
        - soup: beautifulsoup object of stenogram
         - steno_link: str of url to steno
    -OUTPUT:
        - list of conversations info (steno content + agenda)"""
    conversations=get_agenda(soup, steno_link)
    conversations=get_main_talk(soup, conversations, steno_link)
    return conversations

def get_links_stenos(links):
    """function to get all steno content from links
    -INPUT:
        - links: list of links to stenos
    -OUTPUT:
        -dict of {steno_link:steno_content}"""
    stenos={}
    for i, link in enumerate(links):
        if i%20==0:
            print(f'working on {i}')
        steno_soup=BeautifulSoup(requests.get(link).content)
        talk=[]
        try:
            talk=parse_steno(steno_soup, link)
        except Exception as e:
            print(e)
            print(link)
        stenos[link]={'talk':talk}
    return stenos

## Collect päevakorrad which have links to stenos

%time paevakorrad2022=get_years_paevakorrad(range(2022, 2024))

%time paevakorrad=get_years_paevakorrad(range(2010, 2023))

len(paevakorrad)

len(paevakorrad2022)

## How many stenos do we have for each year

df_stenos_links2022=paevakorrad_stenos_count(paevakorrad2022)
df_stenos_links2022

df_stenos_links=paevakorrad_stenos_count(paevakorrad)

df_stenos_links

## Get all unique memos

unique_steno_links=list(set([item.replace('hans-frontend.riigikogu.ee', 'stenogrammid.riigikogu.ee')
                             for sublist in df_stenos_links.stenos_links.to_list() for item in sublist]))
len(unique_steno_links)

unique_steno_links2022=list(set([item.replace('hans-frontend.riigikogu.ee', 'stenogrammid.riigikogu.ee')
                             for sublist in df_stenos_links2022.stenos_links.to_list() for item in sublist]))
len(unique_steno_links2022)

%time stenos1=get_links_stenos(unique_steno_links[:900])

%time stenos2=get_links_stenos(unique_steno_links[900:])

stenos1.update(stenos2)

len(stenos1)

%time stenos2022=get_links_stenos(unique_steno_links2022)

## Save

with open('data/raw/stenos_from_links.json', 'w') as f:
    json.dump(stenos1, f)

with open('data/raw/stenos_from_links2022.json', 'w') as f:
    json.dump(stenos2022, f)
