==> 0_3_clean_text.ipynb <==
import re
import pandas as pd
import numpy as np
from pathlib import Path
from collections import Counter

pd.set_option('display.max_colwidth', -1)

## Data

df=pd.read_csv('data/interim/stenos.csv')
df2023=pd.read_csv('data/interim/stenos2023.csv')

df.shape

df2023.shape

df.year.value_counts()

df2023.year.value_counts()

df=df[df.year!=2022].append(df2023)
df.shape

df.head(1)

df.text.isnull().sum()

#remove cases where text is missing - might be some data collection errors,
#but not much compared to overall amount of data (less than 2%)
df=df[~pd.isnull(df.text)]
df.shape

## See chars we have - is there something funny

cnt_char=Counter(''.join(df.text).lower())

len(cnt_char)

cnt_char.most_common(200)

## Make text without punctuation (migth be useful for topic modelling)
## Also remove some junk

def remove_punct(text):
    """remove punctuation, junk from text
    -INPUT:
        -text: str of text to be cleaned
    -OUTPUT:
        -cleaned text str"""
    text=re.sub('(?<=[^0-9])\. ', ' ', text) #keep . if they are between characters (no space) or after numbers
    text_clean=text.lower().replace('\x98', ' ').replace('\x9f', ' ').replace('\x84', ' ').replace('\x9b', ' ')\
    .replace('\xad', ' ')\
    .replace('\xa0', ' ').replace('...', ' ')\
    .replace('!',' ').replace('\?',' ').replace('?',' ').replace(', ',' ').replace(')',' ')\
    .replace('(',' ').replace('"',' ')\
    .replace('ºC', ' kraadi ').replace(' – ', ' ').replace(':', ' ').replace('–', ' ').\
    replace('[---]', ' ').replace('[', '').replace(']', '')\
    .replace(';', ' ').replace('+', ' pluss ').replace(' & ', '&').replace('”', ' ').\
    replace('„', ' ').replace('=', ' võrdub ')\
    .replace('´', "'").replace('’', "'").replace('“', ' ').replace(' ±', ' pluss-miinus ').\
    replace(' ..',' ').replace('`', "'")\
    .replace('…', ' ').replace('″', ' ').replace(' = ', ' võrdub ')
    text_clean=' '.join(text_clean.split())
    if text_clean.endswith('.'):
        text_clean=text_clean[:-1]
    return text_clean

remove_punct('Austatud Riigikogu, tere päevast! Teate mis, ')

df['text_wo_punct']=df.text.apply(remove_punct)

#we should have less funny characters now
cnt_char_clean=Counter(''.join(df.text_wo_punct).lower())
print(len(cnt_char_clean))
cnt_char_clean.most_common(200)

## Remove "v a h e a e g"

#not really said but indicates when vaheaeg started
df['text_wo_punct']=df.text_wo_punct.str.replace('v a h e a e g', '')
df['text_wo_punct']=df.text_wo_punct.str.replace('v a h e a eg', '')

## Save data

df.to_csv('data/interim/stenos_cleaned.csv', index=False)
