==> 0_4_add_factions.ipynb <==
import re
import requests
import json
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup

## Helpers

def get_query_jsn(url):
    """returns json from request
    -INPUT:
        -url: str of url to be requested
    -OUTPUT:
        - list of päevakorrad"""
    return json.loads(requests.get(url).content)

def get_soup(link):
    """returns beautifulsoup object from link content
    -INPUT:
        -link: str of url to be requested
    -OUTPUT:
        - beautifulsoup of link content"""
    soup=BeautifulSoup(requests.get(link).content)
    return soup

#historic faction data is in html
def get_soup_faction_members(soup):
    """function to get historic faction members information
    -INPUT:
        -soup: beautifulsoup ob ject of page where faction member info is
    -OUTPUT:
        -list of faction members information"""
    factions=[]
    faction_name=soup.find('h3').text
    faction_members=[]
    for i, tag in enumerate(soup.find('h3').next_siblings):
        if tag.name=='h3' or tag.name=='h2':
            faction_name=tag.text
            factions.extend(faction_members)
            faction_members=[]
        #esimehed, aseesimehed
        if tag.name=='p':
            text=tag.text
            text=text.replace('Aseesimees: ', '').replace('Esimees: ', '').replace('Aseesimehed: ', '').replace('Esimehed: ', '').replace('\xa0', ' ')
            names=text.split(', ')
            for name in names:
                for subname in name.split('\n'):
                    #usually names not so long
                    if len(subname.split())<10 and subname!='':
                        faction_members.append((subname, faction_name))
        #list of faction members
        if tag.name=='ul' or tag.name=='ol':
            list_members=[(li.text, faction_name) for li in tag.findAll('li')]
            faction_members.extend(list_members)

    factions.extend(faction_members)
    return factions

def get_koosseis_faction_members(koosseisud=['xi', 'xii', 'xiii', 'xiv']):
    """get faction members from Riigikogu koosseisud
    -INPUT:
        -koosseisud:  list of str of koosseis number
    -OUTPIT:
        -pd.DataFrame of faction members information"""
    df_factions=pd.DataFrame()
    for koosseis in koosseisud:
        print(f'working in faction {koosseis}')
        soup=get_soup(f'https://www.riigikogu.ee/tutvustus-ja-ajalugu/riigikogu-ajalugu/{koosseis}-riigikogu-koosseis/fraktsioonid/')
        faction_members=get_soup_faction_members(soup)
        df_=pd.DataFrame(faction_members)
        print(f'about {df_.shape[0]} members on {koosseis}')
        df_.columns=['name', 'faction']
        df_['koosseis']=koosseis
        df_factions=df_factions.append(df_)
    df_factions=df_factions.drop_duplicates()
    df_factions=df_factions.reset_index(drop=True)
    return df_factions

## Get all historic factions from html pages

df_factions=get_koosseis_faction_members()

df_factions.shape

df_factions.faction.value_counts()

df_factions.head()

## Get current koosseis using API

current_koosseis=get_query_jsn('https://api.riigikogu.ee/api/plenary-members?lang=et')

df_current_faction=pd.DataFrame([[i['fullName'], i['factions'][0]['name']] for i in current_koosseis])
df_current_faction.columns=['name', 'faction']
df_current_faction['koosseis']='xv'
df_current_faction.head()

## Append current and historic data

df_factions_all=df_factions.append(df_current_faction)
df_factions_all=df_factions_all.reset_index(drop=True)
df_factions_all.shape

## Do some cleaning

df_factions_all['name'].nunique(), df_factions_all['faction'].nunique()

#a bit of cleaning and formatting
df_factions_all['name']=df_factions_all.name.str.strip()
df_factions_all['faction']=df_factions_all.faction.str.strip()
df_factions_all['name'].nunique(), df_factions_all['faction'].nunique()

## Get times (if from or to exists) if person has changed faction during koosseis

def get_from_to(text):
    """extract from and to in faction member nae field"""
    if ' (alates' in text.lower():
        return text.split(' (alates')[-1].replace(')', '').strip(), None
    start_end=re.findall('\(\d\d\.\d\d\.\d\d\d\d–\d\d\.\d\d\.\d\d\d\d\)', text)
    if len(start_end)>0:
        start, end = start_end[0].split('–')
        return start.replace('(', '').strip(), end.replace(')', '').strip()
    return None, None

df_factions_all['from'], df_factions_all['to']= zip(*df_factions_all['name'].map(get_from_to))

df_factions_all['from'].value_counts()

df_factions_all[df_factions_all.name.str.contains('kaja')]

## Fill members from and to based on koosseis start time

#this was taken from wikipedia
#NB! default time format MM.DD:YYYY
koosseis_start={'xi':'02.04.2007',
                'xii':'04.04.2011',
                'xiii':'30.03.2015',
                'xiv':'04.04.2019',
               'xv':'10.04.2023'}

koosseis_end={'xi':'03.04.2011',
                'xii':'29.03.2015',
                'xiii':'03.04.2019',
                'xiv':'09.04.2023',
              'xv':'09.04.2027'}

koosseis_start={'xi':'04.02.2007',
                'xii':'04.04.2011',
                'xiii':'03.30.2015',
                'xiv':'04.04.2019',
               'xv':'04.10.2023'}

koosseis_end={'xi':'04.03.2011',
                'xii':'03.29.2015',
                'xiii':'04.03.2019',
                'xiv':'04.09.2023',
              'xv':'04.09.2027'}

df_factions_all['from']=np.where(pd.isnull(df_factions_all['from']), df_factions_all.koosseis.apply(lambda x: koosseis_start[x]), df_factions_all['from'])
df_factions_all['to']=np.where(pd.isnull(df_factions_all['to']), df_factions_all.koosseis.apply(lambda x: koosseis_end[x]), df_factions_all['to'])
df_factions_all['from']=pd.to_datetime(df_factions_all['from'])
df_factions_all['to']=pd.to_datetime(df_factions_all['to'])

df_factions_all

## Do additional cleaning of names

#remove time info from names
df_factions_all['name_clean']=df_factions_all.name.apply(lambda x: x.split('(')[0].split(', ')[0].strip())
df_factions_all.head()

df_factions_all.name_clean.value_counts()

#check if have some funny characters
set(''.join(df_factions_all.name_clean.tolist()).lower())

df_factions_all[df_factions_all.name_clean.str.contains('-')]

# members that have multiple factions in one koosseis - became MPs without faction (left faction/party)
df_factions_all[df_factions_all.duplicated(subset=['name_clean', 'koosseis'])]

## Save faction data

df_factions_all.to_csv('data/interim/factions2022.csv',index=False)

df_factions_all.to_excel('data/interim/factions2022.xlsx',index=False)

## NB! Better to use manually reviewed data, there are some mistakes removed

df_factions_all=pd.read_excel('data/interim/factions2022.xlsx')
df_factions_all.shape

df_factions_all=df_factions_all.drop_duplicates()
df_factions_all.shape

## Read stenos to add faction info

df_stenos=pd.read_csv('data/interim/stenos_cleaned.csv')
df_stenos.shape

df_stenos.head()

## Add koosseis info

#create koossseis variable
df_stenos['date']=pd.to_datetime(df_stenos.year.astype(str)+'.'+df_stenos.month.astype(str)+'.'+df_stenos.day.astype(str))
df_stenos['koosseis']=np.where(df_stenos['date']<=pd.to_datetime(koosseis_end['xi']), 'xi', '')
df_stenos['koosseis']=np.where((df_stenos['date']<=pd.to_datetime(koosseis_end['xii']))&(df_stenos['date']>pd.to_datetime(koosseis_end['xi'])), 'xii', df_stenos['koosseis'])
df_stenos['koosseis']=np.where((df_stenos['date']<=pd.to_datetime(koosseis_end['xiii']))&(df_stenos['date']>pd.to_datetime(koosseis_end['xii'])), 'xiii', df_stenos['koosseis'])
df_stenos['koosseis']=np.where((df_stenos['date']<=pd.to_datetime(koosseis_end['xiv']))&(df_stenos['date']>pd.to_datetime(koosseis_end['xiii'])), 'xiv', df_stenos['koosseis'])
df_stenos['koosseis']=np.where((df_stenos['date']<=pd.to_datetime(koosseis_end['xv']))&(df_stenos['date']>pd.to_datetime(koosseis_end['xiv'])), 'xv', df_stenos['koosseis'])

pd.to_datetime('04.10.2023').day

df_stenos.koosseis.value_counts()

df_stenos.shape

df_stenos[df_stenos.year==2023].head(10)

## Add faction info

#this one takes time!
def get_stenos_factions(x):
    """adds faction member info to steno based on the time of Riigikogu koosseis and when somebody was in faction"""
    df_  = df_stenos[df_stenos.index_snd==x['index_snd']]
    if not df_['speaker'].tolist()[0] in df_factions_all.name_clean.tolist():
         return ''
    df_factions_all_sub=df_factions_all[(df_factions_all['to']>=df_.date.tolist()[0])&(df_factions_all['from']<=df_.date.tolist()[0])&(df_factions_all.name_clean==df_['speaker'].tolist()[0])]
    if df_factions_all_sub.shape[0]==0:
        return ''
    elif df_factions_all_sub.shape[0]==1:
        return df_factions_all_sub['faction'].tolist()[0]
    else:
        print(f'df shape {df_.shape} should be (0, n)')
        
%time stenos_factions=df_stenos.apply(get_stenos_factions, axis=1)

df_stenos['fation']=stenos_factions

df_stenos['fation'].value_counts()

## Save stenos

df_stenos.to_csv('data/interim/stenos_cleaned_faction2022.csv', index=False)
