==> 1_1_gender_bias_w2v.ipynb <==
import math
import gensim
import umap
import hdbscan
import numpy as np
import pandas as pd
from ast import literal_eval
from collections import Counter
from sklearn.cluster import AgglomerativeClustering
import matplotlib.pyplot as plt
from src.gender_bias import SC_WEAT, get_word_vecs

## Load wordvectors

vecs_w2v=gensim.models.KeyedVectors.load('wordvectors/riigikogu_word2vec.wordvectors')

np.expand_dims(vecs_w2v['õed'], axis=0).shape

## Get word vectors

#Attribute Words
female_stimuli = ['naise','naine','tüdruk','tüdruku','õde', 'naiste','naistele','naist','õed','tütar', 'naisterahvas', 'tüdrukud', 'tüdrukute', 'ema', 'emad', 'naistel', 'üksikema']
male_stimuli = ['mehe','mees','poiss','poisi', 'vend', 'meeste','meestele','meest','vennad','poeg', 'meesterahvas', 'poisid', 'poiste', 'isa', 'isad', 'meestel', 'üksikisa']

len(set(female_stimuli))==len(set(male_stimuli))

female_embeddings=get_word_vecs(vecs_w2v, female_stimuli)
male_embeddings=get_word_vecs(vecs_w2v, male_stimuli)

female_embeddings.shape, male_embeddings.shape

## Get target words - use frequency and remove stopwords

df=pd.read_csv('data/interim/stenos_cleaned_topics_vecs.csv')
df.shape

df.tokes[0]

with open('utils/estonian-stopwords.txt', encoding='utf-8') as f:
    stopwords=f.read().split('\n')
print(len(stopwords))  
with open('utils/estonian-stopwords_additional.txt', encoding='utf-8') as f:
    stopwords_addit=f.read().split('\n')
    
stopwords.extend(stopwords_addit)
stopwords.extend(['.', ',', '!', '?', ':', ';', ')', '(', '-', ']', '[', '...', '–,'])
stopwords.extend(female_stimuli)
stopwords.extend(male_stimuli)

tokens=df.tokes.apply(literal_eval).tolist()
tokens=[item for sublist in tokens for item in sublist]

len(tokens), len(set(tokens))

%time cnt=Counter(tokens)

cnt['vanapoiss']

MIN_COUNT=20
most_common_tokens_cnt=cnt.most_common(100000)
most_common_tokens_cnt=[el for el in most_common_tokens_cnt if el[1]>=MIN_COUNT]
most_common_tokens=[el[0] for el in most_common_tokens_cnt if el[0] not in stopwords]
most_common_tokens=[el for el in most_common_tokens if el in vecs_w2v]
most_common_tokens[:10]

len(most_common_tokens)

## Calculate SC-WEATS

STEP = 10000
N_ITERS=math.ceil(len(most_common_tokens)/STEP)
PERMUTATIONS = 10000

df_bias=pd.DataFrame()
for i in range(N_ITERS):
    print(i)
    targets = most_common_tokens[i*STEP:(i+1)*STEP]
    bias_array = np.array([SC_WEAT(vecs_w2v[word],
                                   female_embeddings, male_embeddings,PERMUTATIONS) for word in targets])
    bias_df_ = pd.DataFrame(bias_array, columns=['female_effect_size','female_p_value'])
    df_bias=df_bias.append(bias_df_)

df_bias.index=most_common_tokens

df_bias[(df_bias.female_effect_size>.5)&(df_bias.female_p_value<.05)].sort_values('female_effect_size', ascending=False).head(30)

## Save

df_bias.to_csv('data/interim/gender_bias/words_female_scores_v2.csv')

## Load 

df_bias=pd.read_csv('data/interim/gender_bias/words_female_scores_v2.csv', index_col=0)
df_bias.shape

df_bias

df_bias_female=df_bias[(df_bias.female_effect_size>=.5)&(df_bias.female_p_value<=.05)].sort_values('female_effect_size', ascending=False)
df_bias_female.shape

df_bias_female.head(20)

df_bias_male=df_bias[(df_bias.female_effect_size<=-.5)&(df_bias.female_p_value>=.95)].sort_values('female_effect_size', ascending=True)
df_bias_male.shape

df_bias_male.head(20)

## Cluster female and male words

def cosine_similarity(vecs_ar1, vecs_ar2, words=None):
    """returns vecs_ar.shape[0] x vecs_ar.shape[0] dataframe of each word cosine similarity with other words"""
    sims= np.dot(vecs_ar1, vecs_ar2.T)/(np.linalg.norm(vecs_ar1)*np.linalg.norm(vecs_ar2.T))
    df_sims=pd.DataFrame(sims)
    if isinstance(words, list):
        df_sims.index=words
        df_sims.columns=words
    return df_sims

def cluster_sims(df_sims, n_clusters):
    clustering = AgglomerativeClustering(n_clusters=n_clusters, compute_distances=True).fit(df_sims)
    df_clusters=pd.DataFrame({'token':df_sims.index, 'cluster':clustering.labels_})
    df_clusters.index=df_clusters.token
    return clustering, df_clusters

#### word embeddings

df_bias_female_embs=pd.DataFrame(get_word_vecs(vecs_w2v, df_bias_female.index))
df_bias_female_embs.index=df_bias_female.index

df_bias_male_embs=pd.DataFrame(get_word_vecs(vecs_w2v, df_bias_male.index))
df_bias_male_embs.index=df_bias_male.index

#### cosine similarity between vectors

df_bias_female_embs_sims=cosine_similarity(df_bias_female_embs, df_bias_female_embs, list(df_bias_female_embs.index))

df_bias_male_embs_sims=cosine_similarity(df_bias_male_embs, df_bias_male_embs, list(df_bias_male_embs.index))

#### Agglomerative clustering

clust_clf_female_ag, df_female_clusters_ag=cluster_sims(df_bias_female_embs_sims, n_clusters=20)

df_female_clusters_ag.cluster.value_counts()[:20]

clust_clf_male_ag, df_male_clusters_ag=cluster_sims(df_bias_male_embs_sims, n_clusters=20)

#### HDBSCAN

#first reduce dimensions
reducer = umap.UMAP(random_state=42,n_components=4, metric='cosine')
bias_female_embs_sims_red=dding=reducer.fit_transform(df_bias_female_embs_sims)

#first reduce dimensions
reducer = umap.UMAP(random_state=42,n_components=4, metric='cosine')
bias_male_embs_sims_red=dding=reducer.fit_transform(df_bias_male_embs_sims)

bias_female_embs_sims_red.shape, bias_male_embs_sims_red.shape

clust_hdb = hdbscan.HDBSCAN()
clust_hdb.fit(pd.DataFrame(bias_female_embs_sims_red))

#number of clusters
len(set(clust_hdb.labels_))

clust_hdb_male = hdbscan.HDBSCAN()
clust_hdb_male.fit(pd.DataFrame(bias_male_embs_sims_red))

df_female_clusters_ag['cluster_hdb']=clust_hdb.labels_
df_male_clusters_ag['cluster_hdb']=clust_hdb_male.labels_

df_female_clusters_ag.cluster_hdb.value_counts()

df_female_clusters_ag

#add it to original data
df_bias_female['clust_ag']=df_female_clusters_ag.cluster
df_bias_female['cluster_hdb']=df_female_clusters_ag.cluster_hdb

df_bias_male['clust_ag']=df_male_clusters_ag.cluster
df_bias_male['cluster_hdb']=df_male_clusters_ag.cluster_hdb

vecs_w2v.most_similar('poiss', topn=20)

## Female clusters

df_bias_female.groupby('clust_ag')['female_effect_size'].median().sort_values(ascending=False).head(20)

df_bias_female.groupby('cluster_hdb')['female_effect_size'].median().sort_values(ascending=False).head(20)

df_bias_female[df_bias_female.cluster_hdb==-1]

## Male clusters

df_bias_male.groupby('cluster_hdb')['female_effect_size'].median().sort_values(ascending=True).head(20)

## Create 2d vectors for visualization

df_bias_all_embs=df_bias_female_embs.append(df_bias_male_embs)
df_bias_all_embs_sims=cosine_similarity(df_bias_all_embs, df_bias_all_embs, list(df_bias_all_embs.index))

reducer = umap.UMAP(random_state=42,n_components=2, metric='cosine')
bias_all_embs_sims_red=reducer.fit_transform(df_bias_all_embs)

df_bias_all_embs_sims_red=pd.DataFrame(bias_all_embs_sims_red)

df_bias_female['x']=df_bias_all_embs_sims_red[0].tolist()[:df_bias_female.shape[0]]
df_bias_female['y']=df_bias_all_embs_sims_red[1].tolist()[:df_bias_female.shape[0]]

df_bias_male['x']=df_bias_all_embs_sims_red[0].tolist()[df_bias_female.shape[0]:]
df_bias_male['y']=df_bias_all_embs_sims_red[1].tolist()[df_bias_female.shape[0]:]

df_bias_male

## Add counts

df_bias_male['count']=[cnt[x] for x in df_bias_male.index]
df_bias_female['count']=[cnt[x] for x in df_bias_female.index]

## Save

df_bias_male.to_csv('data/interim/gender_bias/separated/bias_male_w2v.csv')

df_bias_female.to_csv('data/interim/gender_bias/separated/bias_female_w2v.csv')
