==> 2.1_create_summarize_with_gpt.ipynb <==
import pandas as pd
import numpy as np
from configparser import ConfigParser
from tqdm import tqdm
import os
import tiktoken
from openai import OpenAI
from configparser import ConfigParser

## Conf

#setup openai client
config=ConfigParser()
config.read('conf/openai.ini')

client = OpenAI(api_key=config['openai']['apikey'])

## Load data

df=pd.read_excel('data/interim/summarization/for_summarization_mbart_2048_chunks.xlsx')
df.shape

## estimate number of tokens and take subset of data for training

#M tokens
df['ntoks_gpt35'].sum()/1000000

df.index_pk.nunique()

indexes2use=df.sample(n=500, random_state=1).index_pk.unique().tolist()
len(indexes2use)

df_sample=df[df.index_pk.isin(indexes2use)]
df_sample=df_sample[df_sample.ntoks_space>80]
df_sample.shape

df_sample.ntoks_space.describe()

df_sample.ntoks_gpt35.sum()/1e6

## example usage

df_sample.text[150]

chunk=df_sample.text[150]
completion = client.chat.completions.create(
  model="gpt-3.5-turbo",
  messages=[
    {"role": "system", "content": "You are a helpful assistant."},
    {"role": "user", "content": f"Summarize this chunk: {chunk} Provide up to 3 short bullet points into the summary. Use only Estonian."}
  ]
)

print(completion.choices[0].message)

print(completion.choices[0].message.content)

## loop and generate summaries

df_sample

def get_repsonse(prompt, text):
    completion = client.chat.completions.create(
          model="gpt-3.5-turbo",
          messages=[
            {"role": "system", "content": "You are a helpful assistant."},
            {"role": "user", "content": prompt}
          ]
        )
    
    text=completion.choices[0].message.content
    ntoks=completion.usage.total_tokens
    return text, ntoks

savepath='data/interim/summarization/for_summarization_mbart_2048_chunks_summaries.xlsx'

def generate_prompt(text):
    prompt=f"""Summarize text. Provide up to 3 short bullet points into the content summary what was said or asked in the text by whom. 
Use only Estonian. Mark bullets with - . Text: {text}"""
    return prompt

df_sample['summary']=''
df_sample['ntoks_used_gpt35']=-1

df_sample=df_sample.reset_index(drop=True)

n_fails=0

start_i=28
i=start_i
for text in tqdm(df_sample.text.tolist()[start_i:]):
    try:
        summary, ntoks=get_repsonse(generate_prompt(text), text)
    except Exception as e:
        print(i)
        print(e)
        summary=''
        ntoks=-1
        n_fails+=1
        if n_fails>5:
            break
    finally:
        df_sample.loc[i, 'summary']=summary
        df_sample.loc[i, 'ntoks_used_gpt35']=ntoks
        i+=1
    if i%10==0:
        df_sample.to_excel(savepath, index=False)

df_sample.head(30)

df_sample.tail(30)

!dir/s data\interim\summarization\for_summarization_mbart_2048_chunks_summaries.xlsx

## Make train, test split

df_sample.index_pk.nunique()

idx_test=df_sample.sample(n=20, random_state=23).index_pk.unique().tolist()
len(idx_test)

df_sample[df_sample.index_pk.isin(idx_test)].shape

df_sample['split']=np.where(df_sample.index_pk.isin(idx_test), 'test', 'train')

## Save

df_sample.to_excel('data/interim/summarization/for_summarization_mbart_2048_chunks_summaries.xlsx')
