==> 3.0_data4scattertext.ipynb <==
import pandas as pd

df=pd.read_csv('data/interim/stenos_cleaned_faction2022.csv')
df.shape

df.head()

df.tail()

df.speaker.value_counts()[:40]

df_subset=df[df.speaker.isin(['Jevgeni Ossinovski', 'Aivar Sõerd'])]
df_subset.shape

df_subset.to_excel('data/scattertext/ossin_soerd.xlsx', index=False)
